+++ src/main/java/io/github/hapjava/server/impl/json/AccessoryController.java
@@ -13,6 +13,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
 import java.util.concurrent.CompletableFuture;
+import java.util.stream.Collectors;
 import javax.json.Json;
 import javax.json.JsonArrayBuilder;
 import javax.json.JsonObject;
@@ -31,12 +32,18 @@ public class AccessoryController {
 
     Map<Integer, List<CompletableFuture<JsonObject>>> accessoryServiceFutures = new HashMap<>();
     for (HomekitAccessory accessory : registry.getAccessories()) {
-      int iid = 0;
       List<CompletableFuture<JsonObject>> serviceFutures = new ArrayList<>();
-      for (Service service : registry.getServices(accessory.getId())) {
-        serviceFutures.add(toJson(service, iid));
-        iid += service.getCharacteristics().size() + 1;
+
+      Map<Integer, Service> servicesByInterfaceId = registry.getServices(accessory.getId());
+
+      Map<Object, Integer> iidLookup = new HashMap<>();
+      iidLookup.putAll(swapKeyAndValue(servicesByInterfaceId));
+      iidLookup.putAll(swapKeyAndValue(registry.getCharacteristics(accessory.getId())));
+
+      for (Service service : servicesByInterfaceId.values()) {
+        serviceFutures.add(toJson(service, iidLookup));
       }
+
       accessoryServiceFutures.put(accessory.getId(), serviceFutures);
     }
 
@@ -64,16 +71,18 @@ public class AccessoryController {
     }
   }
 
-  private CompletableFuture<JsonObject> toJson(Service service, int interfaceId) throws Exception {
+  private CompletableFuture<JsonObject> toJson(Service service, Map<Object, Integer> iidLookup)
+      throws Exception {
     String shortType =
         service.getType().replaceAll("^0*([0-9a-fA-F]+)-0000-1000-8000-0026BB765291$", "$1");
     JsonObjectBuilder builder =
-        Json.createObjectBuilder().add("iid", ++interfaceId).add("type", shortType);
+        Json.createObjectBuilder().add("iid", iidLookup.get(service)).add("type", shortType);
     List<Characteristic> characteristics = service.getCharacteristics();
     Collection<CompletableFuture<JsonObject>> characteristicFutures =
         new ArrayList<>(characteristics.size());
     for (Characteristic characteristic : characteristics) {
-      characteristicFutures.add(characteristic.toJson(++interfaceId));
+      Integer iid = iidLookup.get(characteristic);
+      characteristicFutures.add(characteristic.toJson(iid));
     }
 
     return CompletableFuture.allOf(
@@ -85,7 +94,20 @@ public class AccessoryController {
                   .map(future -> future.join())
                   .forEach(c -> jsonCharacteristics.add(c));
               builder.add("characteristics", jsonCharacteristics);
+
+              if (!service.getLinkedServices().isEmpty()) {
+                JsonArrayBuilder jsonLinkedServices = Json.createArrayBuilder();
+                service.getLinkedServices().stream()
+                    .map(iidLookup::get)
+                    .forEach(jsonLinkedServices::add);
+                builder.add("linked", jsonLinkedServices);
+              }
+
               return builder.build();
             });
+  }
+
+  private <K, V> Map<V, K> swapKeyAndValue(Map<K, V> map) {
+    return map.entrySet().stream().collect(Collectors.toMap(Entry::getValue, Entry::getKey));
   }
 }